Loading and unloading of cogs
